+++ carbonize/calculators/trains.py
@@ -14,7 +14,7 @@ class Train(Calculator):
     def __init__(self, *, distance: Km) -> None:
         self.distance = distance
 
-    def __repr__(self) -> str:  # pragma: no cover
+    def __repr__(self) -> str:
         return f"Train({self.distance} km)"
 
     def get_step(self) -> Step: